requirements.txt
birdy==0.2
oauthlib==2.0.0
pymongo==3.3.1
python-dateutil==2.6.0
requests==2.11.1
requests-oauthlib==0.7.0
six==1.10.0
wheel==0.24.0
